apps/backend/requirements/base.txt
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
psycopg[binary]==3.3.3
